Increase the require js waitSeconds time to 60s.
Increase the waitSeconds time from 7s(the default) to 60s.

The problem was that when the all.js and all-vue.js load
for more than the 7s, Load timeout for module exception is
thrown and it stops loading next files.

Change-Id: I8aa21a988d4d635b4c53eaa257ddf72dc2f0cc8c
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/1883
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>