Introducing elastic placement zones
EPZ is a Compute Service concept, i.e. it is transparent to the
underlying photon model.

Together with the CRUD service, a task is implemented that regularly
checks EPZ definitions and updates participating computes accordingly.
The task is expected to be scheduled to run on a regular basis but
that will be done in a subsequent change.

Change-Id: I74bcb41bd7dd9c728f8f241a65880d52748d3f54
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/1993
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Rostislav Georgiev <rgeorgiev@vmware.com>