Fix CompositionTaskService contract to follow description - state pattern
The CompositionTaskService was accepting a composite component description
link and resulted in a resource links of container, network, etc.
This is also breaking the request broker service in such a way that
for container description we get a link of a container state
for composite description we get links of container,network states.
The result needs to be the link of the top level description instance - composite
component.

Fix composition request to be based on composite component type

Change-Id: I8d6f28ce8c6c37d7f914ca7fb3d9a1e8a7219226
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2000
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>