VBV-549 - Performance: Containers are not properly cleaned up after the unit test execution
The problem was that when admiral is in mock and a test is executed in
cluster when removing a container we remove the container reference from
a map. To remove it we pass the full uri of the container e.g.
host:port/containerId. However in cluster we have multiple nodes and for
the remove operation we can get an address/port different from the one
we have in the map.
- Changed the logic to check for the path in order to remove the
container from the mock
- Added additional assert in the performance unit test to check if the
containers are removed

Change-Id: Ic9e619fca3d0aa234413c21950fb6a22a12318a6
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2003
Reviewed-by: Aleksandar Angelov <aangelov@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>