Fix bug related with the same user names
The problem was that the query returns result
when there is two users with the same names, but from
different tenants.

The solution is to set the occurence of the tenant to
MUST_OCCUR when there is a user set in the tenant links.

Implemented unit tests.

Change-Id: I5b9df7501e423c509f4d193ee455f583c3ebc691
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2094
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>