DockerNetworkAdapterService: take into account all NetworkState properties
Currently, when the DockerNetworkAdapterService creates the network on the
docker host, only the name and driver properties are read from the
NetworkState instance and only they are passed to docker. This change adds
all the other properties (e.g. options, ipam) to the list of properties
that are taken into account.

Change-Id: I876e7032f1552f8d9cd93dfe69b9693f14513d6f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2106
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>