Placement selection tasks now work with elastic placement zones
A utility to work with query-driven resource pools is provided as
part of this change. It will be also used to fix other places
where we query resource pool computes (for example in
ContainerHostDataCollectionService).

Eventually we can contribute this helper to photon-model since
the query-driven resource pool is a photon-model concept.

Change-Id: I2a10ff9671a4830fe3680f9420e0a75bb55c64b1
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2116
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Peter Mitrov <pmitrov@vmware.com>