Add 'containerStateLinks' runtime property to ContainerNetworkState
This is a list of links to all containers that are registered in a given
network. It is populated on network create/inspect and its contents are
based on the contents of the 'containers' runtime property that is part of
the 'network inspect' command.

Change-Id: If18462a283ec76b67188adf9c6820220b6d84e23
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2498
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>