Compress the agent image
Compress agent file when building it to reduce it's size. Current size is
reduced from 155MB to 35MB. That reduces the uber jar size, and the time
need to transfer the file to the docker host.
Update readme documentation adding information that xz is required to
build the agent.

Change-Id: I76699c0badc944240467c8f4e61ec27e822fc904
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3959
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>