[VBV-599] Container stats log excessive exception stack trace
When a exception message is returned from docker to admiral,
we need to check the status code of it. Unfortunattely there
is no way to get the status code, because the exception is
generic.

The only way is to check if the response message contains
404 status code.

Change-Id: I0b46a8422068dc0f4ec2badca30a803fe1905a6b
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2536
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Ilia Pantchev <ipantchev@vmware.com>