Renamed entities and added tests.
Changed the name of Groups to Projects. The urls to the API are the same.

Changed the name of Policy to Placements. The urls to the API are changed
too.

Added first tests for containers, applications, resource pools, placements
and projects.

Entities will not have Print method anymore, they will return output
string which will be formatted and printed from the command. The idea is
printing of everything to be handled by the command, and to be easier for
testing. In this change this way is used to print Placements only, for the
other entities it will be in the next change. Columns will be dynamically
aligned to prevent broken output.

Change-Id: I4e170c8f1bc2c914636b8251505e23caf0c69360
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2579
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>