[VBV-595] Error messages not hidden
Error messages stays persistent there even after
going back and forth between pages.

The problem was that the error is still in the model
after going to the different view and then go back to
the view with the message.

The solution is to clear the error when the current view
is opening.

Change-Id: I46ff37223d452c11fe7e641d91f4fa9e10b41eb3
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2592
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>