[UI]Added "Containers" link for each network in the network list
In every network item in the existing networks list now there is a
"Containers" link that displays the number of containers that are
currently connected to this network. Pressing this link will
display only the containers that are connected to this network. Containers
are queried by the name of the network.

Change-Id: Ib2bd2cb9748afb1543a4264aa892bf1b93d17c3e
VIDEO: http://url.vmware.com/ue7c
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2643
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>