ContainerClusteringTaskServiceTest.testContainerClusteringTaskServiceIncrementByOneWithDisabledHost
Fix https://jira-hzn.eng.vmware.com/browse/VBV-617, the problem is that we do not wait for ComputeState
to be patched and patch it again. From time to time the two patches are executed it inverce order which
fails the state check in the end.

Change-Id: Iaeed3cfae0f2f0d006b9f6270b269bb41a19b187
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2650
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>