Updated RequestBroker network provisioning logic
Until now a request to RequestBroker as PROVISION_RESOURCE for a network was
just doing the allocation, since that's the way it's being called from the
Composition tasks, where the Composition orchestrates all the allocations
first and then all the actual provisionings second.

The problem is that for provisioning external networks the request to
RequestBroker as PROVISION_RESOURCE should do both parts (allocation +
provisioning) directly since it does not rely on the Composition tasks. To
achieve that we make use of the same "flag" ("__allocation_request" = true)
that the Composition passes to tell the RequestBroker to stop after the
allocation. That's the same flag that the ContainerAllocationTaskService
uses to know whether doing the actual provisioning when requesting a single
container without going through the Composition.

Based on those changes, refactored and simplified the tests from
WordpressProvisioningIT so issues from VBV-624 should be fixed.

Change-Id: Ie2f8905bce9433f9f80320012d35a7dad003400c
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2708
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>