[VBV-601] Http Session expiration in Admiral in vRA mode
The problem was that when you are not using the
embedded Admiral for a long time, then every request
returns 401. The shinding proxy we use has a
token which is generated at the start up of the
application and its default timeout is 1 hour.
(https://opengrok.eng.vmware.com/source/xref/vcac-git/core/
shell/ui/app/src/main/resources/config/shell-default-container.js)

The solution is to check for 401 and to reload the
current page, which in this case will redirect to the
login page.

TODO: Validate with the authentication enabled.
Validate the token is updating when calling vRA.
Change-Id: Ie3f7ade4e92800679002c3f149e574207d28211d
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2753
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>