Fix tests failing with "shell container not available"
The problem was that for some reason while executing all the tests in
the job sometimes the data collection for a given host remains locked. A
lot of the tests are using the same id when adding a host and this is
causing the issue. 
- Added unlock for the data collection before adding a new host

This change should fix the failing tests. However the issue with the
locked data collection should be investigated.

Change-Id: I54724dbb5991e859dd0938d9accd5e48d9e485fd
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2786
Reviewed-by: Jose Dillet <jdillet@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>