[VBV-601] Http Session expiration in Admiral in vRA mode (Fix)
The problem was that when you are not using the
embedded Admiral for a long time, then every request
returns 401.

The solution is to check for 401 and to reload the
current page, which in this case will redirect to the
login page.

Tested in standalone and embedded.

Change-Id: Id5877d60e5dbe19293dc591cbdd69a4a294039f5
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2805
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
Reviewed-by: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>