[UI] Add power state and number of parent hosts for the networks in the network list
The 'Network' header for each item in the networks list is now replaced
with the current state of the network (e.g. CONNECTED, PROVISIONING, etc.)

Also, the number of hosts this network is available on is displayed. This
could be made clickable if information for available networks gets stored
in the compute state.

Changes to the backend have been made in a previous commit.

SCREENSHOT: http://url.vmware.com/pzpc

Change-Id: I955a3856e67196bcfde208919789aeab6ed34bf7
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2797
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>