[VBV-631] The provisioning process of a networked application appears stuck
The provisioning process of a networked application appears stuck
in the UI when provisioning fails. The problem is that when we remove
the network we make a call to get the compute in which the network is placed,
but while the network state is in state PROVISIONING, the originatingHostLink
is not still set and it is null. Then we make a call to the compute
with no compute link and it throws Timeout exception.

The solution is to check if the network state has this originatingHostLink,
and if it doesn't have it to remove the network without going through
the adapter request. As a result the request will fail with the main reason
for failiure.
Change-Id: Ie65a4bab449f11c008a92e2318851bb16b3e3075
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2811
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Stanislav Hadjiiski <hadjiiskis@vmware.com>