Networking UI fix - correct css styles applied
A previous commit have changed the class of the header for networks in the
network list from "container-header" to "on-header". This commit brings
back the correct "container-header" class.

THe changes can be seen on the screenshot bellow: wrong styles for the
left network and correct styles for the right one.

SCREENSHOT: http://url.vmware.com/zfpc

Change-Id: I3b57d8e5c2351e4480b60dd0273546f9db877d2a
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2831
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>