Bump photon-model to 0.4.18 + task for updating placement capacity
The idea is to use this new task for both containers and compute, and
have it scheduled to run periodically as well as explicitly when needed
(e.g. on user actions such as Add Host when we want to show immediate
placement update to the user). This will be done in a subsequent change.

For compute (more specifically vSphere), the usage figures (available
memory, cpu usage %) depend on collected stats which are not yet
enabled (but implemented in the photon model adapter).

Change-Id: I340012319792e3fae6acc3204fb6778d83a7bf15
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2849
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Rostislav Georgiev <rgeorgiev@vmware.com>