[VBV-644][UI] Network list looks strange
Before this change, for networks with connected containers no remove
button was shown (because such networks cannot be deleted on docker
hosts). However, it was not that clear why button was missing. This commit
brings the remove button back for all networks, but pressing it will
produce an error message (displayed in alert box).

Also, before this change it was not possible to delete networks states
that were in retired state because the ContainerNetworkRemovalTaskService
would fail to delete the network on the actual host (since there was no
such network already). Some logic was added to the aforementioned task
service to only delete the network state of retired networks.

Additionally, there was a problem with the styles of the icons of the
networks in the networks list that was fixed in a previous commit.

VIDEO: http://url.vmware.com/d6pc

Change-Id: I015e4950ae0597c54f6e6591b38e385b06cd1363
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2887
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>