VBV-652 - Data Collection: Missing __Containers custom property
The problem happens when we have the same host added in 2 different
tenants. The issue could happen only in CAFE mode. The problem was that
when adding the host to the first tenant the system container is
provisioned. When adding the same host to another tenant we discover the
system container but when we try to create the container state we get
"service already started" because we already have this system container
state. 
- Added check for "Service already started" exception in order not to
fail the data collection when we try to create the system container

Change-Id: I68f6bb4b93f3067f4a08386f7940cc0184b9d8c6
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2944
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>