Reorder affinity filters:
The issue is with a wordpress application with service links. When you provision the
application all containers go to the same host as expected. However, when you try to
increase the count of the wordpress instance provisioning fails because of no hosts
available, but it should put the new container on the same host. This happens because
the ClusterAntiAffinityFilter goes first and excludes the host where all containers are
as there are other hosts avaialable, then the ServiceLinkAffinityFilter rules out all
the rest and we end up with no hosts to choose from.

The fix is to reorder the filters and give precedence to ServiceLinkAffinityHostFilter

Bug Number:
Documentation Notes:
PrecheckinTest URL: <required>
QA Notes:
Review URL: <required>
Reviewed by: <required>
Testing Done: <required>

Change-Id: Ia9b37801f67b3cc16a675c51d3b8c6508ee92d83
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/2954
Bellevue-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
Reviewed-by: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>
Tested-by: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>