Service links getting messed up with multiple hosts and cluster size > 1
The issue is reproduced with a wordpress application where we have 2 hosts and both wordpress and mysql
have cluster size 2. What happens is that after the two mysql containers get provisioned on the two hosts
the wordpress containers are mapped to the wrong mysqls and provisioning fails

The issue is that we use the same DescName object in each HostSelection. When we add a
continer name to the DescName each HostSelection that contains this DescName gets it too. The fix is
to clone the DescName before modifying it

Note that after this fix the result of the provisioning will be basically two separate instances of
wordpress placed on the two hosts because the service links don't work between the two hosts

Bug Number:
Documentation Notes:
PrecheckinTest URL: <required>
QA Notes:
Review URL: <required>
Reviewed by: <required>
Testing Done: <required>

Change-Id: Ic6b8fe5808de89057acd908e3127216c45be3e2f
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3015
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>
CS-Verified: e_vcoauto_glob_1 <e_vcoauto_glob_1@vmware.com>