VBV-666: Error on deleting multiple Applications with scaled containers
When a container is scaled and there's a request to remove all of them,
there could be a race condition that makes all the container removal tasks
to try to remove the same container description, and that shouldn't fail
with an exception.

Added unit test to validate it.

Change-Id: I7164023390ecba16d80684fe492fb453f9176993
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3026
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Igor Stoyanov <istoyanov@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>