Support for https connections
First of all, the CLI http client now loads the system cert pool and
user's trusted certificates which are being stored inside ~/.admiral-cli,
next to the configuration and token file.

If the Admiral instance which the CLI is currently working against,
provide certificate which can't be found in the http client's pool of
trusted certificates, the CLI prompts the user if he wants to trust the
certificate, in case the user agree, the new certificate is being added to
the file holding user's trusted certificates, and the request is being
resend.

The other case is when there arise error which is again x509 related but
the reason is not because the certificate is self signed. Then the user is
again prompted to trust the certificate, and if the user agree, the
certificate is again added to this file. Then these certs from the file
are being loaded in the memory too, and on next requests the CLI will
check if the faulty certificate is already accepted from the user, it will
simply toggle the http client's property "InsecureSkipVerify" to true.

With this change all issues related to x509 certificates should be
handled properly.

Change-Id: I7b5e6dc95ff516b43802c45e472d77a56451b90d
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3174
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>