[VBV-682] The number shown in a network for containers is incorrect
The issue was observed when at least two hosts with configured overlay
were connected to Admiral. On data collection Admiral sends an inspect
command and retrieves a list of the connected containers. Up until now, it
was expected that the output always contains the IDs of the containers, so
they were used to build a list of links to the coresponding
ContainerStates and in the UI the length of this list was shown. However,
when the network is in overlay mode, the IDs will be returned only for
those containers that are physically located on the same host that the
network inspect command was issued, for the rest of the containers an
endpoint ID will be returned (which differs from the ID of the container).
To be able to collect the endpoint IDs for a given container, it is needed
to bump the used API level to 1.21+. To avoid this, in this change the
list of the links to container states was exchanged for the number of the
connected containers (which can be retrieved by network inspect command on
any host). There actually is no information lost, because we have the
opposite relation (the containers know which networks they are connected
to).

Change-Id: I2a0fa4314ab679c8ae373f5d522f34be32cfa95c
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3177
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>