Create only the default-placement when embedded
- "embedded" property creates a defautlt-placement and
default-placement-zone based on its value

- While starting the xenon with the embedded mode, there is
an exception in the console. The problem is that we create
a default compute-description named "docker-host-compute-desc-id"
and then in the ContainerHostDataCollection we use the
ResourcePoolQueryHelper which is used for simplify the retrieval
of computes per placement-zone. When there is no placement-zone,
it fails.

Change-Id: Ib8685d60051badfeacfc6fef4b25c774ca6afa13
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3223
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>