VBV-670: Periodic enumeration of EPZ per compute
Finding the EPZ per compute is expensive with the query-driven EPZs.
For the purposes of the UI, computes per EPZ are periodically enumerated
and the EPZs found are stored in compute's customProperties.

The placement logic (and the entire backend) does not use this
cached value -- it continues to directly use the query for real-time
behavior.

On the UI only a single EPZ is shown per host/compute in order
to minimize the changes. Showing multiple can be added in a
subsequent change.

Change-Id: I575c9cd1242ea95cf062372360f319a3a339d5a5
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3275
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Rostislav Hristov <rhristov@vmware.com>