Fix placement issue when provisioning using an existing network
The problem was that when provisioning an application that is using an
external network the host for provisioning was picked randomly and it
was possible to try to provision on a host that does not have the
network. This was causing the provisioning to fail

- Added checks in case of existing network to check the host containing
on this network and select it for the provisioning
- The fix should work both for bridge and overlay. In case of bridge the
host with the network will be returned.
- In case of overlay the cluster containing the network will be
returned.

Change-Id: Iff3b6d6e9ca476f70d6f1cf31d89cd9b15bb5b13
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3323
Reviewed-by: Zahari Ivanov <zaharii@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>