Compute integration test fixes
Still not fully working but wanted to submit the initial fixes.

1. Rely on the default endpoint resource pool instead of creating a new one.
Using a custom RP for the endpoint is not a clean solution at the moment -
you first have to create the RP and then create the endpoint with this RP
(and this will trigger enumeration into this RP). The problem is that the
RP has no link to the endpoint and the enumeration fails.
2. Various error message fixes.
3. Fix the failure when placement logic finds more computes than the
requested resource count.
4. Added support for a custom enumeration RP in the EndpointAdapterService
(have used it in one version but currently it is not being used).

Change-Id: I29dbe679fe03e8c229efb3a0a882e1a0e66400fd
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3391
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Rostislav Georgiev <rgeorgiev@vmware.com>