[VBV-739] External Network appears as Provisioning when the request has failed
Currently, The NetworkRemovalTaskService is trying to first delete the
network on the host and only then to remove the network state. With this
change on cleanup removals will again first try to remove the actual
network but will remove the network state even if the previous action
failed.

Change-Id: Ie1b25c9435571a1e67b32084936093908a149ed3
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3392
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>