VBV-736 - System container is not updated from 0.5.1 to 0.9.1 as expected
The problem was that if an old system container exists first we discover
it, create the container state and after that delete it and provision a
new version. When provisioning we already had the container state from
the discovered container but the state was missing the volumes.
- Added the volumes even when we create the state for the system
container when discovered.
- Added additional assert to check this scenario

Change-Id: I9f7d347d6f7fa480541839cd57744423b9fb6e01
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3515
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Lazarin Lazarov <llazarov@vmware.com>