[VBV-725] Disable access to the static files outsude vRA
Problem: When in embedded mode, it was possible to access
the static files directly.

Solution: Check in every resource service, whether the xenon
is in embedded mode and also check for specific header which
comes from the cafe.container.ui proxy.
Needs to be checked in the jenkins env.

Change-Id: I0d40c02b9ae98df5a63c592dbd1f4f644eb8c068
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3531
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>