Missing links when deploying clustered application
There are several problems here:
* when deploying an application with a clustered container linked by
another container in the aplication, the link is added only to one of
the containers in the cluster.
* looking at the code, it seems the above or similar should be
reproducible with networks and volumes. Still have to investigate
* If said application is deployed and the cluster is scaled again, the
new links are not added/removed. Not sure about networks and volumes.

This review addresses only the first issue as a starting point. The
others will follow.

Change-Id: Ib06089b4703514056b01e899240303b77a1fc5a4
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3578
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>