VBV-744: Exporting a composite component duplicates dependsOn field.
When exporting to ComponentTemplate, there is a field dependsOn which is filled
from the dependsOn field on the description. Then, dependsOn field is in the
ComponentTemplate and in the description. When exporting to composite component,
the dependsOn in the description should be removed because it shouldn't be
exported to the template.

Change-Id: I86a025844194a4ad353d30aaba80938cf1ac0d07
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3550
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>