VBV-759 - Delete container fails in cluster of 3 nodes when one of them is down
The problem was that when sending the request to the removal task
service there a counter subtask was created with header
.addRequestHeader(Operation.REPLICATION_QUORUM_HEADER,
Operation.REPLICATION_QUORUM_HEADER_VALUE_ALL). This should wait until
the entity is created on all of the nodes and it works as expected when
all of the nodes are up and running. However when one of the nodes is
down this fails with "connection refused".
- Opened an issue in xenon -
https://www.pivotaltracker.com/story/show/133242025
- Removed the headers

Change-Id: Ic6f1b9c45d9a5d4dec3889d444d1ebac4e371861
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3575
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Zahari Ivanov <zaharii@vmware.com>