[VBV-758] After scaling a container the page is not refreshed
This was caused by a recent change in the backend - the
ContainerClusteringTaskService was replaced by a generalized
ClusteringTaskService that handles clustering of multiple resource types.
This change also changed the displayed name of the task from "Container
Clustering" to "Resource Clustering". This commit applies the change in
the UI and brings back the automatic data renewal when the request is
completed.

Change-Id: I096c8f74f802ab40be205156f3b3da9ef59c7bb7
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3586
Reviewed-by: Iveta Ilieva <iilieva@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>