Small updates on networking
- Updated the logic to cast and handle the connectedContainersCount update.
Using getBodyRaw() and casting it could be problematic when Xenon does
cluster communication and body is not the expected object but a plain
String. Using getBody() instead does a valid casting to the expected object
and it's only required to check for possible null values.
- Updated the affinity filter to avoid a possible issue when a host with the
external network that the container(s) require is left out of the selected
hosts in case there are some other hosts set up in a cluster configuration.

Change-Id: I9ef0d17a94c88a8301de52da03f0ffc02ec317e1
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3591
Reviewed-by: Georgi Muleshkov <gmuleshkov@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>