VBV-756 - Empty application after failed provisioning
The problem was that when patching a CompositeComponent for deletion, the
final check whether it could be deleted or not was considering only if the
list of component links was empty or not. Well, in case of composite
components with external networks, the links to those external networks were
still there (those networks aren't deleted at that point) and the
application was not deleted. That was from the Container service with the
calls from the Composition service, in standalone mode it was OK already.

Change-Id: Ie1ad7d9d133a795ea94055a83302162593520876
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3614
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>