VBV-767 - Provisioning an application fails in cluster
The problem was that when we get composite description during
provisioning in the operation we do a query in order to expand the
components side the object this components could be container state,
networks state and so on. In order to handle all the types we use we use
ServiceDocument. The actual problem is that when we get the information
from other node some of the information could not be transferred and we
can get ServiceDocument instead of the original object. The issue
happens only in cluster. 
- Moved the logic for getting component descriptions in the composition
task service. Like this we will always get the real object.

Change-Id: Ib1d35c8a13c22f231781530647eb22ab1dc81a66
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3645
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>