VBV-779: Allow a host to be explicitly assigned to an EPZ
This is to alleviate a possible user confusion when the UI allows
a host to be assigned to an EPZ (tag-based) but the backend ignores
this if the host tags do not match the EPZ tags.

The change is isolated into the tab-based query generation only.

Change-Id: Ib81d4b3f694335013c18d879f6ef24ed98b894c3
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3658
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Rostislav Georgiev <rgeorgiev@vmware.com>