[VBV-773] Scale out container action does not cleanup container state if it fails (at least when connecting the container to a network)
This issue could be hit also on every operation that includes container
provisioning and failing to connect the newly created container to a
network. The problem was that the container state id, status and power
state were only updated if the container was either successfully connected
to all of its networks or has no network connections declared. If the
network connect operation failed for some reason, the container stays in
'allocation' status and PROVISIONING power state with an unknown id.
Because of these, the ContainerRemovalTaskService that was meant to clean
everything up after the unsuccessful provisionig is not able to remove the
container that was created on the host and only removes the corresponding
container state. However, on the next data collection the container that
was created on the host is discovered and added to the list.

This commit adds logic to update the container state with the id of the
created container and sets error status and power state for the container
state.

Additionally, an if-else block in the RequestBrokerService was simplified

Change-Id: I9a3534ad149064595a285116a7be969fcab0ec35
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3682
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>