VBV-776: Support for dashes in PZ self links
When performing a filter by PZ link (e.g. if the PZ link is clicked
on the Hosts view, or when counting hosts per PZ on the UI), the
created OData query contains something like
  customProperties.__epz_default-placement-zone eq 'true'
which breaks the OData handling in cafe container and the result is
always 0 hosts.
The problem only occurs in cafe mode for default-placement-zone.
Also old-style xenon links (guids) could experience the same issue.

The fix is to change the query to:
  customProperties eq '__epz_default-placement-zone'
(which acts as 'contains' for map keys) where dashes in the
value are fully supported.

Manually tested in standalone mode as well as in cafe mode.

Change-Id: Ib95340e56121e10f02347e50aa855e701f3c33c8
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3686
Reviewed-by: Tony Georgiev <tgeorgiev@vmware.com>
Reviewed-by: Rostislav Georgiev <rgeorgiev@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Bellevue-Verified: jenkins <jenkins@vmware.com>