Update on VBV-773 + test
Added some logic to the resolution of VBV-773 that prevents a FAILED task
to be transitioned to FINISHED state when the created container is
inspected.

Also added a test that verifies that failing to connect a newly created
container to a network will not result in skipping data collection for the
container. Previously, skipping this step lead to creating containers with
unknown IDs that couldn't be removed from the docker host.

Change-Id: I04427b6bb1c7bc86182461d405c95240bbd602ee
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3705
Bellevue-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>