VBV-788: Fix issue with number of connected containers to a network in WordpressProvisioningIT
WordpressProvisioningIT is failing sometimes with the stack trace:

java.lang.AssertionError
	at com.vmware.admiral.test.integration.WordpressProvisioningIT.validateAfterStart(WordpressProvisioningIT.java:317)

The problem is that it finds more connected containers than expected. This
can happen when some containers are connected to a network, and at the same
time there is a network inspect operation (from the network maintenance
service).

A container is connected to a network in 2 steps: connect the
container, then increment the value of connectedContainersCount. However, if
the container is connected and then there is a network maintenance, it is
possible that the network inspect already sees the container connected
(updating connectedContainersCount to the real value) and then the increment
puts a wrong value there. If this happens, for some time (until next maintenance)
the value of connectedContainersCount is bigger than it should be.

To address it, in the ContainerNetworkMaintenance, lastInspectMaintainanceInMicros
is now initialized to the current time, so the first maintenance happens after
5 minutes (there is already a network inspect just after the network is created,
so doing another one for the first maintenance is pointless).

Additionally, in ContainerNetworkMaintenance there was a path where the maintenance
operation was not completed, so adding op.complete() there.

Change-Id: Iec7740619ea1f3281569e6ad96d64200d3bbd149
Reviewed-on: http://bellevue-ci.eng.vmware.com:8080/3716
Bellevue-Verified: jenkins <jenkins@vmware.com>
CS-Verified: jenkins <jenkins@vmware.com>
Reviewed-by: Sergio Sanchez <sanchezs@vmware.com>